Rename ticket server to data server in storage to avoid confusion
