+++ htsget-search/src/htsget/bam_search.rs
@@ -193,7 +193,7 @@ pub(crate) mod tests {
   };
   use crate::htsget::{Class, Class::Body, Class::Header, Headers, Response, Url};
   use crate::storage::local::LocalStorage;
-  use crate::storage::ticket_server::HttpTicketFormatter;
+  use crate::storage::data_server::HttpTicketFormatter;
 
   use super::*;
