+++ htsget-search/src/htsget/bcf_search.rs
@@ -156,7 +156,7 @@ mod tests {
   };
   use crate::htsget::{Class, Headers, Response, Url};
   use crate::storage::local::LocalStorage;
-  use crate::storage::ticket_server::HttpTicketFormatter;
+  use crate::storage::data_server::HttpTicketFormatter;
 
   use super::*;
